Support tvrank.json files
The `tvrank.json` file can be put under the toplevel directory of a title and can contain
e.g. the title ID for exact identification. This is useful when there are multiple matches
for a single title name and year.

The format of the `tvrank.json` file is as follows:

```json
{
    "imdb": {
        "id": "ttXXXXXXXX"
    }
}
```

where "ttXXXXXXXX" is the IMDB title id